requirements_raspi.txt
pandas == 1.2.3
django == 3.1.6
django-pandas == 0.6.4
djangorestframework == 3.12.4
djangorestframework-simplejwt == 4.7.2
django-filter == 2.4.0
gunicorn == 20.0.4
plotly == 4.14.3
psycopg2-binary == 2.8.6
pillow == 8.3.1
requests == 2.26.0